general improvement to build process; cmake libs now use INTERFACE to pull in include dirs, defines and dependend libs; a bit of file structure cleanup;
