+++ CMakeLists.txt
@@ -5,14 +5,15 @@ cmake_minimum_required(VERSION 2.8)
 OPTION(OPTION_BUILD_TESTS "Build Test Projects" ON)
 
 message (STATUS "adding application: grabsend")
-add_subdirectory (grabsend)
+add_subdirectory(grabsend)
+add_subdirectory(grabsend ${CMAKE_BINARY_DIR}/grabsend2)
 
 if (APPLE)
-    add_definitions (-DMAC_OSX -DUNIX)
+    add_definitions(-DMAC_OSX -DUNIX)
     set (MAC_OSX TRUE)
 endif()
 
 if (OPTION_BUILD_TESTS)
-	message (STATUS "adding tests: libgrabbertest")
-	add_subdirectory (tests/libgrabbertest)
+	message (STATUS "adding tests...")
+	add_subdirectory(tests/libgrabbertest)
 endif ()
