+++ libgrabber/CMakeLists.txt
@@ -6,12 +6,47 @@ set (CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${CMAKE_CURRENT_SOURCE_DIR}/../cmak
 include (Common)
 include (FindAllExternals)
 
-add_dzlib()
-add_boost()
-add_libcommon()
+use_lib(libcommon ..)
 
-file (GLOB_RECURSE src_files RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} src/*.cpp)
-file (GLOB_RECURSE header_files RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} src/*.h)
+list(APPEND hdr_dir "${CMAKE_CURRENT_SOURCE_DIR}/..")
+
+set(src "")
+list(APPEND src
+	"src/IDesktopTools.cpp"
+	"src/IDesktopTools.h"
+	"src/grabber_null.cpp"
+	"src/grabber_null.h"
+	)
+
+if(WIN32)
+	list(APPEND src
+		"src/window_info_win32.cpp"
+		"src/bitbltgrabber_win32.cpp"
+		"src/bitbltgrabber_win32.h"
+		"src/DesktopTools_Win32.cpp"
+		"src/DesktopTools_Win32.h"
+		"src/WindowGrabber_Win32.cpp"
+		"src/WindowGrabber_Win32.h"
+		"src/grabber_win32.cpp"
+		"src/grabber_win32.h"
+		"src/screen_enumerator_win32.cpp"
+		"src/screen_enumerator_win32.h"
+		"src/process_info_win32.cpp"
+		)
+endif()
+
+if(APPLE)
+	list(APPEND src
+		"src/osx/GrabberOSX.cpp"
+		"src/osx/GrabberOSX.h"
+		"src/osx/GrabberOSX_mouse.mm"
+		"src/process_info_osx.cpp"
+		"src/window_info_osx.cpp"
+		)
+endif()
+
+#file (GLOB_RECURSE src_files RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} src/*.cpp)
+#file (GLOB_RECURSE header_files RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} src/*.h)
 
 if (MAC_OSX)
 	if (MAC_OSX_VERSION GREATER 106)
@@ -22,11 +57,9 @@ else()
 	set (objcpp_src_files "")
 endif()
 
-include_directories (${HDRS})
-add_library (libgrabber ${src_files} ${header_files} ${objcpp_src_files})
-target_link_libraries(libgrabber ${LIBS})
-
-# if (LINUX)
-# 	# otherwise we get strange linking errors
-# 	target_link_libraries (grabber ${SCREENGRAB_LIBS})
-# endif()
+add_library(libgrabber ${src} ${objcpp_src_files})
+target_link_libraries(libgrabber dzlib libcommon)
+set_target_properties(libgrabber PROPERTIES
+	INCLUDE_DIRECTORIES "${hdr_dir}"
+	INTERFACE_INCLUDE_DIRECTORIES "${hdr_dir}"
+	)
