+++ libgrabber/src/grabber_null.cpp
@@ -1,4 +1,4 @@
-#include "NullGrabber.h"
+#include "grabber_null.h"
 #include <assert.h>
 #include <string.h>
 
@@ -20,23 +20,6 @@ namespace dz
 		if (screen == 0) return Rect (0,0, 1280, 800);
 		if (screen == 1) return Rect (1280, 0, 1920, 1080);
 		return Rect ();
-	}
-
-	void fillWithColor (Buffer * destination, int32_t color)
-	{
-		for (int y = 0; y < destination->height; y++)
-		{
-			int32_t * line = (int32_t*) (destination->data + y * destination->rowLength);
-
-			for (int x = 0; x < destination->width; x++)
-			{
-				uint32_t rintensity = rand() % 255;
-				uint32_t gintensity = rand() % 255;
-				uint32_t bintensity = rand() % 255;
-				uint32_t intensity = (0xff << 24) | (rintensity << 16) | (gintensity << 8) | (bintensity);
-				*(line + x) = intensity;
-			}
-		}
 	}
 
 	VideoFrameHandle NullGrabber::GrabVideoFrame()
