+++ libgrabber/src/screen_enumerator_win32.h
