+++ libgrabber/src/window_info_linux.cpp
